issue #175: rename linux to unix
