+++ internal/pkg/unix/unix.go
@@ -1,4 +1,4 @@
-package linux
+package unix
 
 import (
 	"bufio"
